delete workflow and flag approval, ID HARDCODED!

+++ frontend/src/apiCalls.js
@@ -96,6 +96,10 @@ export function getIndividualTemplateWorkflow(id) {
   return axiosClient.get(`/api/v1/workflow/${id}`);  
 }
 
+export function getIndividualAssignedWorkflow(id) {
+  return axiosClient.get(`/api/v1/workflow/assigned/${id}`);  
+}
+
 export function updateIndividualTemplateWorkflow(data) {
   return axiosClient.put("/api/v1/workflow", JSON.stringify(data));  
 }
@@ -129,8 +133,12 @@ export function createWorkflowAssigned(data) {
   return axiosClient.post("/api/v1/workflow/assigned", JSON.stringify(data));  
 }
 
-export function deleteWorkflow(id) { // unique ID
-  return axiosClient.delete(`/api/v1/workflow/${id}`);  
+export function deleteWorkflow(id, userId) { // unique ID
+  return axiosClient.delete(`/api/v1/workflow/delete/${id}/${userId}`);  
+}
+
+export function deleteAssignedWorkflow(id, userId) { // unique ID
+  return axiosClient.delete(`/api/v1/workflow/assigned/delete/${id}/${userId}`);  
 }
 
 // QUESTIONNAIRE MANAGEMENT
